// cms/src/app/enum/navigation-path-enum.ts
export enum NavigationPathEnum {
  HOME = '/',

  LOGIN = 'auth',
  RESET_PASSWORD = 'auth/reset-password',

  ADMIN_LIST = 'admin-user/list',
  ADMIN_MANAGE = 'admin-user/manage',
  ADMIN_CHANGE_PASSWORD = 'admin-user/change-password',
  ADMIN_EDIT_PROFILE = 'admin-user/edit-profile',
  ADMIN_SETTING = 'admin-user/setting',

  APP_USER_LIST = 'app-user/list',
  APP_USER_MANAGE = 'app-user/manage',
  APP_USER_DEVICE_LIST = 'app-user/device-list',
  APP_USER_LOGIN_HISTORY = 'app-user/login-history',

  CURRENT_AFFAIR_LIST = 'current-affair/current-affair-list',
  CURRENT_AFFAIR_MANAGE = 'current-affair/current-affair-manage',

  EVENT_LIST = 'event/event-list',
  EVENT_MANAGE = 'event/event-manage',
  EVENT_DETAIL = 'event/event-detail',

  FAMILY_LIST = 'family/family-list',
  FAMILY_MANAGE = 'family/family-manage',
  FAMILY_DETAIL = 'family/family-detail',
  FAMILY_EDIT_PERSONAL_DETAIL = 'family/edit-personal-detail',

  JOB_LIST = 'job/job-list',
  JOB_MANAGE = 'job/job-manage',

  MATRIMONIAL_LIST = 'matrimonial/matrimonial-list',
  MATRIMONIAL_MANAGE = 'matrimonial/matrimonial-manage',

  TEMPLE_LIST = 'temple/temple-list',
  TEMPLE_MANAGE = 'temple/temple-manage',

  REPORT_INQUIRY = 'report/inquiry-report',
  REPORT_PUSH_NOTIFICATION = 'report/push-notification-report',

  BUSINESS_LIST = 'lov/business-list',
  BUSINESS_MANAGE = 'lov/business-manage',
  SERVICE_LIST = 'lov/service-list',
  SERVICE_MANAGE = 'lov/service-manage',

  COUNTRY_LIST = 'lov/country-list',
  COUNTRY_MANAGE = 'lov/country-manage',
  STATE_LIST = 'lov/state-list',
  STATE_MANAGE = 'lov/state-manage',
  DISTRICT_LIST = 'lov/district-list',
  DISTRICT_MANAGE = 'lov/district-manage',
  CITY_VILLAGE_LIST = 'lov/city-village-list',
  CITY_VILLAGE_MANAGE = 'lov/city-village-manage',

  GENDER_LIST = 'lov/gender-list',
  GENDER_MANAGE = 'lov/gender-manage',

  ADDICTION_LIST = 'lov/addiction-list',
  ADDICTION_MANAGE = 'lov/addiction-manage',

  RELATION_SHIP_LIST = 'lov/relationship-list',
  RELATION_SHIP_MANAGE = 'lov/relationship-manage',

  RELIGION_LIST = 'lov/religion-list',
  RELIGION_MANAGE = 'lov/religion-manage',

  GOTRA_LIST = 'lov/gotra-list',
  GOTRA_MANAGE = 'lov/gotra-manage',

  POST_LIST = 'lov/post-list',
  POST_MANAGE = 'lov/post-manage',

  JOB_TYPE_LIST = 'lov/job-type-list',
  JOB_TYPE_MANAGE = 'lov/job-type-manage',
  JOB_CATEGORY_LIST = 'lov/job-category-list',
  JOB_CATEGORY_MANAGE = 'lov/job-category-manage',
  JOB_SUB_CATEGORY_LIST = 'lov/job-sub-category-list',
  JOB_SUB_CATEGORY_MANAGE = 'lov/job-sub-category-manage',

  FAQ_CATEGORY_LIST = 'lov/faq-category-list',
  FAQ_CATEGORY_MANAGE = 'lov/faq-category-manage',

  LEGAL_PAGE_LIST = 'lov/legal-page-list',
  LEGAL_PAGE_MANAGE = 'lov/legal-page-manage',

  CONFIG_PARAMETER_LIST = 'misc/config-parameter-list',
  CONFIG_PARAMETER_MANAGE = 'misc/config-parameter-manage',

  FAQ_LIST = 'faq-list',
  FAQ_MANAGE = 'faq-manage',

  CONFIG_PARAM_LIST = 'config-param-list',
  CONFIG_PARAM_MANAGE = 'config-param-manage',

  MANDAL_LIST = 'mandal/mandal-list',
  MANDAL_MANAGE = 'mandal/mandal-manage',
  MANDAL_MEMBER_MANAGE = 'mandal/manage-mandal-member',

  TRUSTEE_LIST = 'mandal/trustee-list',
  TRUSTEE_MANAGE = 'mandal/trustee-manage'
}